// Zybach.Web/src/app/shared/generated/enum/custom-rich-text-type-enum.ts
//  IMPORTANT:
//  This file is generated. Your changes will be lost.
//  Source Table: [dbo].[CustomRichTextType]

export enum CustomRichTextTypeEnum {
  PlatformOverview = 1,
  Disclaimer = 2,
  Homepage = 3,
  Help = 4,
  LabelsAndDefinitionsList = 5,
  Training = 6,
  RobustReviewScenario = 7,
  ReportsList = 8,
  Chemigation = 9,
  NDEEChemicalsReport = 10,
  ChemigationPermitReport = 11,
  ChemigationInspections = 12,
  WaterQualityInspections = 13,
  WaterLevelInspections = 14,
  WellRegistrationIDChangeHelpText = 15,
  ClearinghouseReport = 16,
  SensorList = 17,
  WaterQualityReport = 18,
  AnomalyReportList = 19,
  WaterLevelExplorerMap = 20,
  WaterLevelExplorerMapDisclaimer = 21,
  IrrigationUnitIndex = 22,
  OpenETIntegration = 23,
  SupportTicketIndex = 24,
  WellPumpingSummary = 25,
  WellGroupList = 26,
  WellGroupEdit = 27,
  WaterLevelsReport = 28,
  SensorHealthCheck = 29,
  SensorStatusMap = 30,
  FarmingPractices = 31,
  FlowTestReport = 32
}